requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
Pillow==8.3.1
protobuf==3.17.3
python-dotenv==0.18.0
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
WTForms==2.3.3
